marking web socket handlers as a separate generic piece. PR 1 of 2 (#1232)
